requirements.txt
github3.py==0.9.4
gnureadline==6.3.3
ipython==3.2.0
pep8==1.6.2
python-application==1.5.0
pytz==2015.4
requests==2.7.0
uritemplate.py==0.3.0
wheel==0.24.0
lda==1.0.2
influxdb==2.6.0
Jinja2==2.8
mandrill==1.0.57
babel==2.1.1
schedule==0.3.2
